New models organization
Well, I realized, that DB models and program models should be the same, so made 2 folders inside Models folder, made clases partial and put into one pure db model parts in onother put everything ehat left
